+++ cpp_code/cpp_classes/SideDef.cpp
@@ -5,12 +5,12 @@ SideDef::SideDef(uint16_t x, uint16_t y, uint16_t sectorNumber, std::string uppe
     {}
 std::ostream& operator<<(std::ostream& os, const SideDef& obj) {
     os << "SideDef{ ";
-    os << "x: " << obj.x << " ";
-    os << "y: " << obj.y << " ";
+    os << "x: " << (int) obj.x << " ";
+    os << "y: " << (int) obj.y << " ";
     os << "upperTextureName: " << obj.upperTextureName << " ";
     os << "lowerTextureName: " << obj.lowerTextureName << " ";
     os << "middleTextureName: " << obj.middleTextureName << " ";
-    os << "sectorNumber: " << obj.sectorNumber << " ";
+    os << "sectorNumber: " << (int) obj.sectorNumber << " ";
     os << "}";
     return os;
 }
