+++ cpp_code/cpp_classes/SubSector.cpp
@@ -6,8 +6,8 @@ SubSector::SubSector(uint16_t segCount, uint16_t firstSegNumber)
 
 std::ostream& operator<<(std::ostream& os, const SubSector& obj) {
         os << "SubSector{ ";
-        os << "segCount: " << obj.segCount << " ";
-        os << "firstSegNumber: " << obj.firstSegNumber << " ";
+        os << "segCount: " << (int) obj.segCount << " ";
+        os << "firstSegNumber: " << (int) obj.firstSegNumber << " ";
         os << "}";
         return os;
     }
