+++ cpp_code/cpp_classes/Thing.cpp
@@ -19,11 +19,11 @@ Thing::Thing() {
 
 std::ostream& operator<<(std::ostream& os, const Thing& obj) {
     os << "Thing{ ";
-    os << "x: " << obj.x << " ";
-    os << "y: " << obj.y << " ";
-    os << "angle: " << obj.angle << " ";
-    os << "doomType: " << obj.doomType << " ";
-    os << "flags: " << obj.flags << " ";
+    os << "x: " << (int) obj.x << " ";
+    os << "y: " << (int) obj.y << " ";
+    os << "angle: " << (int) obj.angle << " ";
+    os << "doomType: " << (int) obj.doomType << " ";
+    os << "flags: " << (int) obj.flags << " ";
     os << "skillLevel12: " << obj.skillLevel12 << " ";
     os << "skillLevel3: " << obj.skillLevel3 << " ";
     os << "skillLevel45: " << obj.skillLevel45 << " ";
